+++ CHANGES.rst
@@ -1,13 +1,16 @@
 Changelog
 =========
 
+
+
 0.1a2
 -----
 
-- Added settings for the grid.
+* Added settings for the grid.
+* Move initialization of the grid to the template.
 
 
 0.1a1
 -----
 
--   Initial release.
+* Initial release.