+++ CHANGES.rst
@@ -1,13 +1,12 @@
 Changelog
 =========
 
-
-
 0.1a2
 -----
 
 * Added settings for the grid.
 * Move initialization of the grid to the template.
+* Added some documentation in the README.
 
 
 0.1a1